requirements.txt
# Core dependencies
python-telegram-bot>=21.0
python-dotenv==1.0.0

# Web server (for webhook mode)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Async support
httpx>=0.24.0

# Database - Async SQLAlchemy & PostgreSQL
sqlalchemy>=2.0.20
alembic>=1.13.0
psycopg[binary]>=3.2.0
aiosqlite>=0.19.0

# Fuzzy matching for debtor search
thefuzz>=0.20.0
python-Levenshtein>=0.21.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
